Add User model with registration functionality
This commit introduces the 'app/models/User.php' file, which contains the User class. The User class includes a constructor for setting up a database connection and a 'register' method for creating new user accounts.

In addition to the 'register' method, the User class also has two private helper methods:
1. 'generateUUID': Generates a UUID for each new user.
2. 'hashPassword': Takes a plain-text password and returns a hashed version of it using front and back salts, as well as the SHA-256 hashing algorithm.

With these methods in place, the User model can be used to handle user registration in the Taskmaster application.