Add login method and update register method in User model
Added a new method called "login" to the User model which takes an email and password and returns the user object if the email and password match a record in the database. Updated the "register" method to check if an email is already registered before inserting a new record to the database. Also added two new methods "isEmailRegistered" and "getUserByEmail" to check if an email is already registered and to get the user object by email respectively.