Add login process functionality
Added login process functionality to the website by creating a new file called login_process.php in the public folder. The file is responsible for receiving login form data through the POST method, authenticating the user by calling the login() method of the AuthController class, and redirecting to the dashboard page if the user is successfully authenticated. The file also displays an error message if the authentication fails.