Add logout functionality to index.php
In this commit, I added the functionality to logout from the website by creating a new case in the switch statement in index.php. The new case 'logout' destroys the session and redirects the user to the index.php page. I also made changes to the default case, which now leads to a welcome page with options to either register or login. This is a significant improvement to the website's authentication system, as users can now log out of their accounts.