Add welcome.php in app/views/auth/ folder
Add welcome.php file to the app/views/auth/ folder which will serve as the default page for the application. This page displays options to either log in or register. The commit summary reflects the addition of the welcome.php file to the folder, and the commit description outlines the contents of the file.