Add dashboard page and Add email registration validation
Added dashboard.php in the app/views directory to display the dashboard page for the authenticated user. This page will only be displayed if the user is authenticated, otherwise it will redirect to the login page.

Modified register_process.php in the public directory to check if the email has been registered before inserting new data to the database. If the email has been registered, the registration process will fail and an error message will be displayed.